\name{Rfast-package}
\alias{Rfast-package}
\docType{package}
\title{
Really fast R functions
}

\description{
A collection of Rfast functions for data analysis.
}

\details{
\tabular{ll}{
Package: \tab Rfast\cr
Type: \tab Package\cr
Version: \tab 1.1 \cr
Date: \tab 2016-07-29\cr
License: \tab GPL-2\cr
}
}

\author{
Manos Papadakis <papadakm95@gmail.com>, Michail Tsagris <mtsagris@yahoo.gr> and Ioannis Tsamardinos <tsamard@csd.uoc.gr> and 
}

\section{Maintainers }{
Manos Papadakis <papadakm95@gmail.com>
}

\note{
Acknowledgments:

We would like to acknowledge Professor Kurt Hornik for his invaluable help with this R package.
}

\keyword{ Mahalanobis distance }
\keyword{ covariance matrix }
\keyword{ variances }
\keyword{ t-tests }
\keyword{ univariate regressions }


