%[dont read]

\name{Rfast-package}
\alias{Rfast-package}
\docType{package}
\title{
\packageTitle{Rfast}
}
\description{
\packageDescription{Rfast}
}

\details{
\tabular{ll}{
Package: \tab \packageName{Rfast}\cr
Type: \tab \packageType{Rfast}\cr
Version: \tab \packageVersion{Rfast} \cr
Date: \tab \packageDate{Rfast}\cr
License: \tab \packageLicense{Rfast}\cr
}
}

\author{
\itemize{
\item Manos Papadakis <papadakm95@gmail.com>
\item Michail Tsagris <mtsagris@uoc.gr>
\item Marios Dimitriadis <kmdimitriadis@gmail.com>
\item Stefanos Fafalios <stefanosfafalios@gmail.com>
\item Ioannis Tsamardinos <tsamard@csd.uoc.gr>
\item Matteo Fasiolo <matteo.fasiolo@gmail.com>
\item Giorgos Borboudakis <borbudak@gmail.com>
\item John Burkardt <jburkardt@fsu.edu>
\item Changliang Zou <nk.chlzou@gmail.com>
}
}

\section{Maintainers}{
\packageMaintainer{Rfast}
}

\note{
Acknowledgments: We would like to acknowledge:

\itemize{
\item Professor Kurt Hornik, Doctor Uwe Ligges (and the rest of R core team) for 
their invaluable help with this R package. 
\item Erich Studerus for his invaluable comments.
\item Neal Fultz for his suggestions. 
\item Vassilis Vasdekis for his invaluable help with the random effects models. 
\item Marios Dimitriadis' work was funded by the Special Account for Research Funds of the University of Crete, Department of Computer Science. 
\item Phillip Si is greatly acknowledged for his help with the Floyd-Warshal algorithm. 
\item Keefe Murphy for his invaluable help with NEWS file and for his suggestions. 
\item Zacharias Papadovassilakis gave us the inspiration for the memory efficient version of the k-NN algorithm. 
\item Yannis Pantazis explained us how the orhtogonal matching pursuit works. 
\item Achim Zeileis for his help with the quasi Poisson regression models. 
\item Pedro J. Aphalo for finding a bug.
\item Dimitris Kyriakis for finding a bug. 
\item Cyrille Conord for finding a bug.
\item Aaron Robotham for finding a bug.
\item Calvin Pan from the Department of Human Genetics at UCLA found a bug in the function glm_logistic and he is greatly acknowledged for that.
\item Adam Muschielok from Rodenstock GmbH found a bug in the function vmf.mle and he is greatly acknowledged for that.
\item Bret Presnell for detecting and correcting a bug in the function rvmf.
\item Gabriel Hoffman for spotting a mistake in the fuction dirimultinom.mle.
\item Lutz von der Burchard for spotting a bug in the function bic.corfsreg.
}
}

\keyword{ G^2 test }
\keyword{ variances }
\keyword{ t-tests }
\keyword{ univariate regressions }
\keyword{ distance }
\keyword{ energy distance }